src/datagen/sql2ibis/pyproject.toml
[project]
name = "sql2ibis"
version = "0.1.0"
description = "Deterministic SQL→Ibis translation pipeline for iBERT training data generation"
requires-python = ">=3.10"
dependencies = [
    "ibis-framework==9.5.0",
    "sqlglot>=25.0.0",
    "duckdb>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.6.0",
]
bigquery = ["google-cloud-bigquery>=3.0.0"]
snowflake = ["snowflake-connector-python>=3.0.0"]
postgres = ["psycopg2-binary>=2.9.0"]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
sql2ibis = "sql2ibis.cli:app"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
